Update `extract_permissions` to use for backfilling optional_permissions
